added rate exchange, fixed mypy, ruff.

+++ finalProject/pos/core/models/report.py
@@ -7,4 +7,4 @@ class Report(BaseModel):
     shift_id: str
     n_receipts: int
     products: Dict[str, int]
-    revenue: int
+    revenue: float
