+++ finalProject/pos/core/models/sales.py
@@ -3,4 +3,4 @@ from pydantic import BaseModel
 
 class Sales(BaseModel):
     n_receipts: int
-    revenue: int
+    revenue: float
